blockhain/vm -> vm
This commit promotes the vm into its own first-class system. It is
one of the most complex areas of the code, and can be implemented
somewhat separately from everything else. This will make the whole
thing much more understandable and implementable, especially if
parties want to implement the VM as a separate, sandboxed process.